app.js: Separate GA related code into component/ga.js
